use 1 file for AnthropicPassthroughLoggingHandler
